### 2019-06-30 #1534 @whedon
[PRE REVIEW]: Blimpy: Breakthrough Listen I/O Methods for Python
**Submitting author:** @telegraphic (<a href="http://orcid.org/0000-0003-2783-1608">Danny Price</a>)
**Repository:** <a href="https://github.com/UCBerkeleySETI/blimpy" target ="_blank">https://github.com/UCBerkeleySETI/blimpy</a>
**Version:** 1.3.6
