### 2019-06-30 #1534 @kyleniemeyer
Hey @arfon, I see `paper.md` and `paper.bib` right at the top level of the repo, any idea why whedon doesn't see those?
