### 2019-06-30 #1534 @kyleniemeyer
Hi @telegraphic, thanks for your submission. While we get things initiated here, I noticed that your article is quite short; although we don't publish full-length papers here, we do expect 250-1000 words (one or two pages in PDF format is typical). Can you expand on your software a bit more, for example talking about features, infrastructure, etc., or even giving one or two examples?
