### 2019-07-03 #1534 @xuanxu
:wave: @garrettj403 and @dhuppenkothen: Would you be willing to review this paper? I think your expertise in Python and Jupyter notebooks and your background on Astronomy/Radioastronomy makes both of you a great fit for this submission.
