### 2019-07-03 #1534 @dhuppenkothen
@xuanxu Unfortunately I'm going to be travelling/teaching/organizing at a school for the next few weeks, so I don't think I can do this on a timescale earlier than about a month from now. I'm really sorry!
