### 2019-07-04 #1534 @garrettj403
Hi @xuanxu, I'm happy to review this repo. I'm not very familiar with SETI or the Breakthrough Listen program, but it looks like an interesting project!
