### 2019-07-05 #1534 @xuanxu
:wave: @benjaminrose — would you be willing to contribute a review for JOSS on this submission?
