### 2019-07-05 #1534 @garrettj403
Hi @telegraphic - In case you didn't know this, ApJ has an agreement with JOSS where you can submit two papers in parallel: one to JOSS and one to ApJ. You can find more info here: [https://iopscience.iop.org/article/10.3847/1538-4357/aaf876](url)
