### 2019-07-05 565c108
verify that files were downloaded
