+++ tests/README.md
@@ -1,0 +1,7 @@
+### Test Data
+
+------
+To remove unwanted bulk from this repository,
+the test data will not be kept in the repo, but will be stored instead on a public server.
+
+Currently, they are [here](http://blpd0.ssl.berkeley.edu/Voyager_data/).
