### 2019-07-06 3886cf8
add guppi tests
