+++ tests/data.py
@@ -4,3 +4,5 @@ here = path.dirname(path.abspath(__file__))
 
 voyager_fil = path.join(here, 'test_data/Voyager1.single_coarse.fine_res.fil')
 voyager_h5 = path.join(here, 'test_data/Voyager1.single_coarse.fine_res.h5')
+voyager_raw = path.join(here, 'test_data/blc3_2bit_guppi_57396_VOYAGER1_0006.0013.raw')
+voyager_block1 = path.join(here, 'test_data/Voyager1_block1.npy')
