### 2019-07-08 #1534 @xuanxu
:wave: @zhampel — would you be willing to contribute a review for JOSS on this submission?
