### 2019-07-09 #1554 @whedon
[REVIEW]: Blimpy: Breakthrough Listen I/O Methods for Python
**Submitting author:** @telegraphic (<a href="http://orcid.org/0000-0003-2783-1608">Danny Price</a>)
**Repository:** <a href="https://github.com/UCBerkeleySETI/blimpy" target ="_blank">https://github.com/UCBerkeleySETI/blimpy</a>
**Version:** 1.4.1
