### 2019-07-18 c8656bc
add offset for 0 values
Former-commit-id: ac96ac450b0afc29fc3abf2ba4f2077d76892f17
