### 2019-07-22 #1554 @zhampel
@xuanxu, @telegraphic, there are a couple of GitHub users (beyond the listed authors) who have contributed to the repo. Specifically FX196 & marksbrt.

@xuanxu are there clear guidelines on defining contributing authors. I just want to check prior to checking some boxes.
