### 2019-07-22 #1554 @xuanxu
@zhampel authorship is a complex topic, here are some guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html#authorship

*the authors themselves are responsible for deciding which contributions are sufficient for co-authorship. Your job as a reviewer is to check that the list of authors appears reasonable, and if it’s not obviously complete/correct, to raise this as a question during the review.*
