### 2019-07-27 c16c9db
Update .travis.yml
Former-commit-id: 68df9b1336613c70dc9415b2816f836b4d351242
