### 2019-07-27 d095cfd
Update .travis.yml
Former-commit-id: d298e33e4e760c1b4c5a8f927edb0ab05a68fb66
