### 2019-07-30 #1554 @garrettj403
Hi, sorry for the delay -- I've been really busy for the last two weeks, but I have a lot more time now. I started by taking a look at the documentation, and I'll try out the software over the next few days.

JOSS paper
----------
- I have the same question as @zhampel about authorship. In the ``setup.py`` file, there are 4 authors listed (Danny Price, Emilio Enriquez, Griffin Foster, Greg Hellbourg), and on GitHub, there are another two authors who seem to have made substantial commits (Yuhong Chen and Mark Siebert).
- The paper is still very short. I think @kyleniemeyer  brought this up in the pre-review, but it hasn't been addressed so far. It would be nice if you could elaborate a bit more (e.g., what exactly you're looking for in the data, what is a "waterfall" file, etc.).
- The references don't display correctly in the [PDF proof](https://github.com/openjournals/joss-papers/blob/joss.01554/joss.01554/10.21105.joss.01554.pdf). This might be an issue with the formatting of the ``paper.bib`` file.
- The ``Croft:2016`` reference has a DOI listed online that is missing from the ``paper.bib`` file (10.13140/RG.2.2.10561.20326). Also, this is not a published ``article``. It should be listed as a ``techreport`` or ``unpublished``, and include a URL to the paper. ``researchgate.net`` also shouldn't be listed as the journal name.
- I wasn't able to find a software archive for ``Blimpy``.
- There are several typos in ``paper.md``:
   - avaiable -> available
   - guppi -> GUPPI
   - astromers -> astronomers
   - endeavour -> endeavor

Documentation
-------------
- The [project API](https://blimpy.readthedocs.io/en/latest/modules.html) is very difficult to read/navigate. Most functions and methods are missing argument descriptions.
- The examples are very limited. The only examples that I could find are in ``README.md``, but they only really show how to import data.
- In ``README.md``, it says "This readme is far from complete. If you have any request/questions please lets us know!". The REAME should be expanded.
- Overall, I'm finding it very difficult to begin using the software with the current level of documentation.
