### 2019-07-30 #1554 @garrettj403
Functionality
-------------
- You should mention in the documentation that you need to use ``git-lfs`` to clone the repository.
- I was able to run the tests found in ``blimpy/tests/``, but the code coverage is currently only 44%. It would be good if this was a bit higher. (Note: If you have code that doesn't require testing, you can flag it using ``# pragma: no cover``. Then it won't be included in coverage calculations.)

Currently, I'm not sure what else I can do with ``Blimpy``. I might need to wait until there are more examples to get me started.
