### 2019-08-05 4325dfb
Update docker_guide.md
Former-commit-id: 9f908247785d26996111c5a4424e8fb6b06a61c8
