### 2019-08-09 #1554 @zhampel
@telegraphic

After installing via `pip install blimpy` and running `python setup.py test`, I receive an error regarding the lack of module `bitshuffle`. Any guidance on properly prepping & running the test suite?

Regarding the docs, I'm not sure the 'Writing Docs' page is necessary nor particularly instructive for the purposes of contributing to the blimpy codebase. I think a contributing guide similar to [this](https://jrbourbeau.github.io/pyunfold/contributing.html) is more like what this requirement is for JOSS publication. I also agree with @garrettj403 that the API is rather unclear.

I think until further example usage and more in depth documentation is provided, I can't provide further review of your submission.
