### 2019-08-10 #1554 @garrettj403
@zhampel

They have a couple extra packages listed under ``extras_require`` in ``setup.py``, including bitshuffle. From your local copy of blimpy, you can run: ``pip install -e .[full]``

Again, this should be explained in the installation instructions.
