### 2019-08-24 0e0407f
Now using six for Py2/Py3 methodtype compatibility
