### 2019-08-24 1520453
Post-merge updates for Py3 + unit tests
Unit tests can now be run from my laptop and pass


Former-commit-id: b38e96d68c7b1174eb2831e4c31c30a90dd7532a
