### 2019-08-24 b38e96d
Post-merge updates for Py3 + unit tests
Unit tests can now be run from my laptop and pass
