+++ setup.py
@@ -28,6 +28,7 @@ install_requires = [
         'cython',
         'h5py',
         'scipy',
+        'six'
 ]
 
 extras_require = {
@@ -46,7 +47,7 @@ setup(name='blimpy',
       install_requires=install_requires,
       extras_require=extras_require,
       url='https://github.com/ucberkeleyseti/blimpy',
-      author='Danny Price, Emilio Enriquez, Griffin Foster, Greg Hellbourg',
+      author='Danny Price, Emilio Enriquez, Yuhong Chen, Mark Siebert, and BL contributors',
       author_email='dancpr@berkeley.edu',
       entry_points=entry_points,
       packages=find_packages(),
@@ -63,5 +64,5 @@ setup(name='blimpy',
       ],
       setup_requires=['pytest-runner'],
       tests_require=['pytest'],
-      test_suite="tests",
+      test_suite="blimpytests",
 )
