### 2019-08-24 2a064ce
Update paper.md
Fix some typos and hopefully the citation syntax.
