### 2019-08-24 0401c74
Update paper.bib with numpy, h5py, astropy citations
Former-commit-id: 001d7f566a95994163b936523d52bc533fcd0b11
