### 2019-08-24 494ce25
Update paper.md
Added a new paragraph with citations to numpy/astropy/matplotlib

Former-commit-id: 9a9146e688a51bdff33308c2ad1eddd5a7de74a1
