+++ paper.bib
@@ -141,16 +141,22 @@ archivePrefix = "arXiv",
 }
 
 
-
-@techreport{Croft:2016,
-author = {Croft, S and Siemion, A and MacMahon, D and Lebofsky, M and Hickish, Jack and Price, D C and Deboer, D and Werthimer, D and Gajjar, Vishal and Isaacson, H},
-title = {{Breakthrough Listen Follow-up of a Transient Signal from the RATAN-600 Telescope in the Direction of HD 164595}},
-journal = {researchgate.net
-},
-year = 2016,
-doi = {0.13140/RG.2.2.10561.20326},
-url = {https://www.researchgate.net/publication/328345994_Breakthrough_Listen_Follow-up_of_a_Transient_Signal_from_the_RATAN-600_Telescope_in_the_Direction_of_HD_164595},
-month = aug
+@INPROCEEDINGS{Croft:2016,
+       author = {{Croft}, Steve and {Siemion}, Andrew and {De Boer}, David and
+         {Enriquez}, J. Emilio and {Foster}, Griffin and {Gajjar}, Vishal and
+         {Hellbourg}, Greg and {Hickish}, Jack and {Isaacson}, Howard and
+         {Lebofsky}, Matt and {MacMahon}, David and {Price}, Daniel and
+         {Werthimer}, Dan},
+        title = "{The Breakthrough Listen Search for Intelligent Life}",
+    booktitle = {American Astronomical Society Meeting Abstracts \#231},
+         year = "2018",
+       series = {American Astronomical Society Meeting Abstracts},
+       volume = {231},
+        month = "Jan",
+          eid = {231.01},
+        pages = {231.01},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018AAS...23123101C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @ARTICLE{Numpy:2011,
