### 2019-08-24 9f59029
Update paper.md
Adding turboseti and other use cases

Former-commit-id: 164f4c8e1067ae0499de19ceab1afc985beb97a5
