+++ paper.md
@@ -38,7 +38,7 @@ has become the largest SETI endeavour in history. This has also resulted in an u
 publicly-available data [@Lebofsky:2019]. Over 1 PB of from the Breakthrough Listen program may be downloaded from
 [seti.berkeley.edu/opendata](http://seti.berkeley.edu/opendata).
 
-The ``Blimpy``--Breakthrough Listen I/O Methods for Python--package provides Python 2.7+/3.6+ utilities
+The ``Blimpy``---Breakthrough Listen I/O Methods for Python---package provides Python 2.7+/3.6+ utilities
 for viewing and interacting with the data formats used within the Breakthrough Listen program.
 This includes Sigproc filterbank (.fil) and HDF5 (.h5) files that contain dynamic spectra (aka 'waterfalls'),
 and GUPPI raw (.raw) files that contain voltage-level data. Python methods for data extraction,
@@ -46,15 +46,15 @@ calibration, and visualization are provided. A suite of command-line utilities a
 
 The waterfall data product stores an array of detected power across frequency channel (i.e. spectra) over time. 
 These files can be several GB in size, with millions of channels and/or time steps. ``Blimpy`` provides 
-convenient methods to extract frequencies and time slices of interest--without loading the full file into memory-- 
-which are presented as ``Numpy`` arrays [@Numpy:2011]. Methods for manipulating lower-level voltage data products
+convenient methods to extract frequencies and time slices of interest---without loading the full file into memory--which 
+are presented as ``Numpy`` arrays [@Numpy:2011]. Methods for manipulating lower-level voltage data products
 stored in the GUPPI raw format, as generated by the Green Bank Telescope, are also provided.
 ``Blimpy`` uses the Matplotlib library [@Pylab:2007] to provide plotting of spectra, time series, and dynamic spectra;
 the ``Astropy`` package for handling of astronomical coordinates [@Astropy:2013; @Astropy:2018]; and, 
-the ``H5py`` package to interact with data stored in HDF5 files [@H5py:2013]. 
+the ``H5py`` package to interact with data stored in HDF5 files [@H5py:2013]. The [turboSETI](https://github.com/UCBerkeleySETI/turbo_seti) pipeline, which conducts doppler acceleration searches for narrowband signals that would indicate the presence of technologically-capable life beyond Earth, uses ``Blimpy`` for file handling and diagnostic plotting.
 
 ``Blimpy`` was designed to be used by radio astronomers, students and anyone else interested in accessing
- Breakthrough Listen data. It has already been used in a number of scientific publications
+ Breakthrough Listen data, whether searching for SETI signals, spectral lines, pulsars, fast radio bursts, or other astrophysical phenomena. It has already been used in a number of scientific publications
 [@Croft:2016; @Enriquez:2017; @Enriquez2018; @Enriquez:2019; @Gajjar:2018; @Price:2019a;  @Price:2019b].
 
 # Acknowledgements
