+++ paper.md
@@ -15,6 +15,8 @@ authors:
     affiliation: "+, 1, 3"
   - name: Yuhong Chen
     affiliation: 2    
+  - name: Mark Siebert
+    affiliation: 2    
 affiliations:
   - name: Equal contribution from both authors
     index: +
