### 2019-08-26 5af3251
Moved gup2hdf to deprecated
Former-commit-id: ea2bdf4b62c97f4986f377a886594a1864eaec8a
