### 2019-09-04 02db693
Update paper.md
Co-Authored-By: Yuhong Chen <mikechen212@gmail.com>
Former-commit-id: 610e90f159a20a4098aa4f8e63fd887d765a6474
