### 2019-09-17 #1554 @telegraphic
Hi @xuanxu, getting there: I'm waiting on a pull request to be reviewed/accepted that addresses a lot of the points raised. (One of the authors just finished their PhD so took some well-deserved time off)
