### 2019-09-25 ce3df4d
Minor edits. Bugfix on citation.
