+++ paper.bib
@@ -115,14 +115,14 @@ archivePrefix = {arXiv},
 
 
 @ARTICLE{Gajjar:2018,
-   author = {{Gajjar}, V. and {Siemion}, A.~P.~V. and {Price}, D.~C. and 
-	{Law}, C.~J. and {Michilli}, D. and {Hessels}, J.~W.~T. and 
-	{Chatterjee}, S. and {Archibald}, A.~M. and {Bower}, G.~C. and 
-	{Brinkman}, C. and {Burke-Spolaor}, S. and {Cordes}, J.~M. and 
-	{Croft}, S. and {Enriquez}, J.~E. and {Foster}, G. and {Gizani}, N. and 
-	{Hellbourg}, G. and {Isaacson}, H. and {Kaspi}, V.~M. and {Lazio}, T.~J.~W. and 
-	{Lebofsky}, M. and {Lynch}, R.~S. and {MacMahon}, D. and {McLaughlin}, M.~A. and 
-	{Ransom}, S.~M. and {Scholz}, P. and {Seymour}, A. and {Spitler}, L.~G. and 
+   author = {{Gajjar}, V. and {Siemion}, A.~P.~V. and {Price}, D.~C. and
+	{Law}, C.~J. and {Michilli}, D. and {Hessels}, J.~W.~T. and
+	{Chatterjee}, S. and {Archibald}, A.~M. and {Bower}, G.~C. and
+	{Brinkman}, C. and {Burke-Spolaor}, S. and {Cordes}, J.~M. and
+	{Croft}, S. and {Enriquez}, J.~E. and {Foster}, G. and {Gizani}, N. and
+	{Hellbourg}, G. and {Isaacson}, H. and {Kaspi}, V.~M. and {Lazio}, T.~J.~W. and
+	{Lebofsky}, M. and {Lynch}, R.~S. and {MacMahon}, D. and {McLaughlin}, M.~A. and
+	{Ransom}, S.~M. and {Scholz}, P. and {Seymour}, A. and {Spitler}, L.~G. and
 	{Tendulkar}, S.~P. and {Werthimer}, D. and {Zhang}, Y.~G.},
     title = "{Highest Frequency Detection of FRB 121102 at 4-8 GHz Using the Breakthrough Listen Digital Backend at the Green Bank Telescope}",
   journal = {\apj},
