### 2019-09-25 4e7b9a2
Bump to 1.3.7. This includes mainly removal of large filterbank files to an outside source, and JOSS paper initial submission (not last version).
Former-commit-id: a6452dce960c7f70dfd2d348c7dd514c4db0c238
