### 2019-09-27 2029dd5
add docker guide to readme
