### 2019-09-29 20e28fc
Updating setup.py for markdown in pypi.org
Former-commit-id: 0c080079e0e7505018c2b193d05eec75e6a9b189
