+++ MANIFEST.in
@@ -1,0 +1,6 @@
+include LICENSE
+include *.md
+include MANIFEST.in
+include requirements*.txt
+include tests
+recursive-include tests *
