### 2019-09-29 0c08007
Updating setup.py for markdown in pypi.org
