+++ setup.cfg
@@ -1,2 +1,5 @@
 [aliases]
 test=pytest
+
+[metadata]
+description-file=README.md
