### 2019-09-29 #1554 @telegraphic
Hmm, whedon is not happy. I just removed a single space in the metadata of `paper.md`, let's see if that fixes it
