### 2019-09-29 #1554 @telegraphic
LGTM Whedon!

@xuanxu, @garrettj403  and @zhampel, thanks for your patience. We have made quite a few changes thanks to your suggestions:

1) Firstly, we have updated the paper.md with references and an expanded author list. The README is also updated, and I've added a link to our [data format paper](https://ui.adsabs.harvard.edu/abs/2019arXiv190607391L/abstract), that is now accepted to PASP (Publications of the Astronomical Society of the Pacific).

2) We removed the need for git-lfs, and now data are downloaded from our self-hosted data archive. We've also moved code coverage over from coveralls over to codecov.io, and worked on improving the coverage to 70%.

3) I added a [jupyter notebook example](https://github.com/UCBerkeleySETI/blimpy/blob/master/examples/voyager.ipynb) into the new `examples` directory, which shows how to use the code to find the telemetry signal from the Voyager telescope from some of our data.

4) We also made a structural change to the package -- you'll notice the code has been rearranged into directories, based on what the code does (as a sidenote, apart from being logical, this makes the coverage reports much more useful!)

BEFORE:

```
blimpy
  |- calib_utils
  |- deprecated
```

AFTER

```
blimpy
  |- calib_utils
  |- deprecated
  |- plotting
  |- ephemeris
  |- io
```

* The `io` directory contains `file_wrapper.py`, the main file handler, `fil_writer.py` and `hdf_writer.py` for writing files, and `sigproc.py` for handling sigproc filterbank files. Anyone wanting to add a new reader or writer need only touch code in this directory.

* The `ephemeris` directory contains `compute_lst.py` for computing sidereal time, and `compute_lsrk.py`, which computes the LSRK for doppler correction of Earth rotation. The lat/lon/elevation of the Parkes and GBT telescopes is included in `config.py` here, so that other telescopes can be added straightforwardly in the future.

* The `plotting` directory houses all the methods for plotting, that were previously methods of the Filterbank/Waterfall classes (they are attached to the class using a create_bound_method function).
