+++ docs/overview.md
@@ -1,34 +1,45 @@
 [![Build Status](https://travis-ci.org/UCBerkeleySETI/blimpy.svg?branch=master)](https://travis-ci.org/UCBerkeleySETI/blimpy)
-[![Coverage Status](https://coveralls.io/repos/github/telegraphic/blimpy/badge.svg?branch=master)](https://coveralls.io/github/telegraphic/blimpy?branch=master)
+[![Documentation Status](https://readthedocs.org/projects/blimpy/badge/?version=latest)](https://blimpy.readthedocs.io/en/latest/?badge=latest)
+[![codecov](https://codecov.io/gh/UCBerkeleySETI/blimpy/branch/master/graph/badge.svg)](https://codecov.io/gh/UCBerkeleySETI/blimpy)
+ [![JOSS status](http://joss.theoj.org/papers/e58ef21f0a924041bf9438fd75f8aed0/status.svg)](http://joss.theoj.org/papers/e58ef21f0a924041bf9438fd75f8aed0)
 
-## Breakthrough Listen I/O Methods for Python
+## Breakthrough Listen I/O Methods for Python.
 
 ### Filterbank + Raw file readers
 
-This repository contains Python 2 readers for interacting with Sigproc filterbank (.fil), HDF5 (.h5) and guppi raw (.raw) files generated
-by the Breakthrough Listen instruments.
+This repository contains Python 2 readers for interacting with [Sigproc filterbank](http://sigproc.sourceforge.net/sigproc.pdf) (.fil), HDF5 (.h5) and [guppi raw](https://baseband.readthedocs.io/en/stable/guppi/) (.raw) files,
+as used in the [Breakthrough Listen](https://seti.berkeley.edu) search for intelligent life.
+
 
 ### Installation
 
-The latest release (1.3.5) can be installed via pip:
+The latest release can be installed via pip:
 
 ```
 pip install blimpy
 ```
 
-Or, the latest version of the development code can be installed from the github [repo](https://github.com/UCBerkeleySETI/blimpy) and then run `python setup.py install` (with sudo if required) or by using the following terminal command:
+Or, the latest version of the development code can be installed from the github [repo](https://github.com/UCBerkeleySETI/blimpy) and then run `python setup.py install` or `pip install .` (with sudo if required), or by using the following terminal command:
 
 ```
 pip install https://github.com/UCBerkeleySETI/blimpy/tarball/master
 ```
 
-You will need numpy, h5py astropy, scipy, and matplotlib as dependencies. A `pip install` should pull in numpy, h5py, and astropy, but you may still need to install scipy and matplotlib separately.
+To install everything required to run the unit tests, run:
 
-Note that h5py needs to be installed in this way.
+```
+pip install -e .[full]
+```
+
+You will need `numpy`, `h5py`, `astropy`, `scipy`, and `matplotlib` as dependencies. A `pip install` should pull in numpy, h5py, and astropy, but you may still need to install scipy and matplotlib separately. To interact with files compressed with [bitshuffle](https://github.com/kiyo-masui/bitshuffle), you'll need the `bitshuffle` package too.
+
+Note that h5py generally needs to be installed in this way:
 
 ```
 $ pip install --no-binary=h5py h5py
 ```
+
+
 
 ### Command line utilities
 
@@ -45,11 +56,11 @@ Use the `-h` flag to any of the above command line utilities to display their av
 
 ### Reading blimpy filterbank files in .fil or .h5 format
 
[…]
