+++ .travis.yml
@@ -15,7 +15,7 @@ before_script:
 - true
 script:
 - pwd
-- docker run -e TRAVIS_JOB_ID="$TRAVIS_JOB_ID" -e TRAVIS_BRANCH="$TRAVIS_BRANCH" -e DIST=${DIST} -e CODECOV_TOKEN=${CODECOV_TOKEN} -it ${DIST}:latest "tests/run_tests.sh"
+- docker run -e TRAVIS_JOB_ID="$TRAVIS_JOB_ID" -e TRAVIS_BRANCH="$TRAVIS_BRANCH" -e DIST=${DIST} -e CODECOV_TOKEN=${CODECOV_TOKEN} -it ${DIST}:latest "ls; code/tests/run_tests.sh"
 
 after_success:
   - docker login -u $DOCKER_USER -p $DOCKER_PASS
