### 2019-10-02 #1554 @garrettj403
Hi @telegraphic

The recent changes look good. It's much easier to understand the functionality with the new changes to the paper and the workflow example.

I still have a few remaining comments that I would like to be addressed:
- I can't see an archive for Blimpy (i.e., through a 3rd party archiving service such as Zenodo)
- The version is 1.3.5 on GitHub, but 1.4.0 on PyPI
- The package API is still a bit of a mess. A lot of the functions are missing descriptions of what they do as well as descriptions of the function arguments.
- The package is missing community guidelines (i.e., "guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support")
- The automated tests currently aren't passing. On Travis, it says that they are passing, but it's not actually collecting any tests (i.e., ``collected 0 items / 1 errors``). It looks like [build 368](https://travis-ci.org/UCBerkeleySETI/blimpy/jobs/592371909) worked correctly. Also, running ``pytest tests/`` on my computer causes it to crash, but that could be because I don't have enough memory.

I'm happy to sign off once these last few issues are fixed (especially the API).
