### 2019-10-03 1685609
fix test script
Former-commit-id: b35f7c061d2f03403be67c60435045aa861df123
