### 2019-10-03 9c038f6
Improved / tidied docstrings
