### 2019-10-03 #1554 @xuanxu
About the Zenodo archive: There is no need to archive the code before the review is finished. That way the archive created will include all changes made during the review process.
