### 2019-10-04 ffadfbc
cd into /code when running docker tests
