### 2019-10-04 48b6f10
Update .travis.yml with new exec cmd for docker
will now run `/bin/bash -c 'cd /code; /code/tests/run_tests.sh'`

Former-commit-id: 43bb335b08c8c33d7c28228fb8d5d4b266171fd8
