### 2019-10-04 #1554 @telegraphic
Hi @garrettj403 and @xuanxu, I've got a draft in Zenodo ready -- it looks like I can fill in the Journal ID as metadata, so I'll fill this in once the review is finished?

We've fixed the Travis-CI problem, added CONTRIBUTIONS.md, and I went through and improved docstrings. I've bumped to v 1.4.1 also, which we will push to PyPi (relevant collaborator is currently on vacation, so might take a few days).
