+++ .travis.yml
@@ -19,8 +19,5 @@ script:
 - docker run -e TRAVIS_JOB_ID="$TRAVIS_JOB_ID" -e TRAVIS_BRANCH="$TRAVIS_BRANCH" -e DIST=${DIST} -e CODECOV_TOKEN=${CODECOV_TOKEN} -it ${DIST}:latest /bin/bash -c 'cd /code; /code/tests/run_tests.sh'
 
 after_success:
-  - docker login -u $DOCKER_USER -p $DOCKER_PASS
-  - export REPO=fx196/blimpy
-  - export TAG=`if [ "$TRAVIS_BRANCH" == "master" ]; then echo "latest"; else echo "$TRAVIS_BRANCH" ; fi`
-  - docker tag ${DIST}:latest $REPO:${DIST%.docker}_$TAG
-  - docker push $REPO
+  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then echo "not PR"; fi
+  - /code/docker/push_to_docker/sh
