+++ .travis.yml
@@ -19,4 +19,4 @@ script:
 - docker run -e TRAVIS_JOB_ID="$TRAVIS_JOB_ID" -e TRAVIS_BRANCH="$TRAVIS_BRANCH" -e DIST=${DIST} -e CODECOV_TOKEN=${CODECOV_TOKEN} -it ${DIST}:latest /bin/bash -c 'cd /code; /code/tests/run_tests.sh'
 
 after_success:
-  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then echo "Not Pull Request, Pushing to Docker Hub"; ls; fi
+  - if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then echo "Not Pull Request, Pushing to Docker Hub"; docker/push_to_docker.sh; fi
