### 2019-10-04 4c1f70a
Remove Docker Hub Pushing for Pull Requests (#110)
* Improved / tidied docstrings

* Create CONTRIBUTING.md

* cd into /code when running docker tests

* Update .travis.yml with new exec cmd for docker

will now run `/bin/bash -c 'cd /code; /code/tests/run_tests.sh'`

* Bump to v1.4.1

* don't push to dockerhub if pr

* don't push if pr

* update yaml

* update yaml

* ls

* fix path
