+++ setup.py
@@ -3,7 +3,7 @@ setup.py -- setup script for use of packages.
 """
 from setuptools import setup, find_packages
 
-__version__ = '1.4.0'
+__version__ = '1.4.1'
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
