### 2019-10-04 #1554 @garrettj403
Hi @telegraphic, the changes to the docstrings look good. For some reason [``blimpy.waterfall.Waterfall``](https://blimpy.readthedocs.io/en/latest/blimpy.html#blimpy.waterfall.Waterfall) doesn't display all of the arguments correctly online, but it does display them correctly using ``pydoc``.

@xuanxu, I'm ready to recommend publication. The documentation is very nice and it includes an [example workflow](https://github.com/UCBerkeleySETI/blimpy/blob/master/examples/voyager.ipynb) to get new users started. The package is easy to install via ``pip``. (It's a bit harder if you want to run the unit tests, but they have included instructions in their ``README`` to tell you how to do that.) Overall, it is an interesting package that allows you to interact with Breakthrough Listen data.
