### 2019-10-09 #1554 @xuanxu
:wave: @zhampel can you review the changes [made by @telegraphic](https://github.com/openjournals/joss-reviews/issues/1554#issuecomment-536267054) and summarize what points (if any) you feel are still remaining? Thanks!
