### 2019-10-14 #1554 @zhampel
@telegraphic

Thanks for the reorganization of the codebase, as it's clearer how the pieces work. The example notebook is a nice addition for usage purposes. Furthermore, thanks for the clarifying the README for streamlined install, testing, usage, and some simple module calls.

@xuanxu I recommend this submission for publication.
