### 2019-10-16 #1554 @xuanxu
@telegraphic: a couple of minor details to fix in the paper:
- from the summary of the paper: `Over 1 PB of from the...`. I guess there's a `data`missing there, so it should be `Over 1 PB of *data* from the...`
- at the end of the summary, when you mention the publications that have already used Blimpy, the `Enriquez et al. 2019` reference is duplicated
