### 2019-10-16 e6d6bcd
Update paper.md
Former-commit-id: 0d2d8c3d1e7006ca9fc0fcfa947ce03bcdce5456
