### 2019-10-16 #1554 @telegraphic
@xuanxu missing 'data' added. I think the bib style is generating 'Price, Croft et al' and 'Price, Enriquez et al', i.e. showing first two authors, to differentiate Price2019a and Price2019b?
