### 2019-10-17 #1554 @xuanxu
Thanks @telegraphic, everything looks good, here are the next steps:

- Please archive the latest release in Zenodo
- Check the Zenodo deposit has the correct metadata: title and authors names should match the paper; you may also add your ORCID.

Once you do that please report here the Zenodo DOI.
