### 2019-10-20 #1554 @xuanxu
Everything looks ready 🎉
Pinging @openjournals/joss-eics for final acceptance
