### 2019-10-20 #1554 @Kevin-Mattheus-Moerman
@telegraphic some minor points about the paper

- [ ] Check your affiliations. The second instance of `Nijmegen` is spelled wrong, it currently says `Nimegen`. Also I believe the correct affiliation there should probably be `Radboud University, Nijmegen, the Netherlands`. Please also add `United States of America`, and spell out `the Netherlands`.
- [ ] In your repo and the Zenodo archive Blimpy is written with a capital B. The paper uses a lower case latter. Can you amend the paper to use a capital letter too? Alternatively change everything to have a lower case letter instead, whatever you prefer.
- [ ] Give your paper a final check again yourself as we'll shortly process formal acceptance.
